+++ public/vocables/modules/tools/KGquery/KGquery_graph.js
@@ -207,6 +207,8 @@ var KGquery_graph = (function () {
      * @memberof module:KGquery_graph
      * @returns {void}
      */
+    // Deprecated method replaced by hypergraphMaker_bot
+    /* 
     self.DrawImportsCommonGraph = function () {
         var source = KGquery.currentSource;
         var sources = [];
@@ -214,8 +216,6 @@ var KGquery_graph = (function () {
         if (imports) {
             sources = sources.concat(imports);
         }
-
-        self.saveVisjsModelGraph();
         var visjsData = { nodes: [], edges: [] };
         var uniqueNodes = {};
         self.KGqueryGraph = new VisjsGraphClass("KGquery_graphDiv", { nodes: [], edges: [] }, self.visjsOptions);
@@ -224,44 +224,34 @@ var KGquery_graph = (function () {
 
             function (source, callbackEach) {
                 var visjsDataSource = { nodes: [], edges: [] };
-                UserDataWidget.listUserData(null, function (err, result) {
-                    if (err) {
+                self.downloadVisjsGraph(source, function (err, result) {
+                    if (err && err != "notFound") {
                         return alert(err || err.responseText);
                     }
-                    // order to get last saved instance of our graph in user_data
-                    result = result.sort((a, b) => new Date(a.created_at) - new Date(b.created_at));
-                    //if graph loaded with loadSaved --> display=checkBox displayGraphInList else last instance graph
-                    var resultId = null;
-                    result.forEach(function (item) {
-                        if (item.data_label == source + "_model") {
-                            resultId = item.id;
-                        }
-                    });
-                    if (resultId) {
-                        UserDataWidget.loadUserDatabyId(resultId, function (err, result) {
-                            if (result && result.data && result.data.data_content) {
-                                visjsDataSource = result.data.data_content;
+
+                    if (result) {
+                        visjsDataSource.nodes = result.nodes;
+                        visjsDataSource.edges = result.edges;
+                    }
+                    if (visjsDataSource.nodes) {
+                        visjsDataSource.nodes.forEach(function (node) {
+                            if (!uniqueNodes[node.id]) {
+                                uniqueNodes[node.id] = 1;
+                                node.x = null;
+                                node.y = null;
+                                //node.fixed = false;
+                                visjsData.nodes.push(node);
                             }
-                            if (!err && visjsDataSource.nodes) {
-                                visjsDataSource.nodes.forEach(function (node) {
-                                    if (!uniqueNodes[node.id]) {
-                                        uniqueNodes[node.id] = 1;
-                                        node.x = null;
-                                        node.y = null;
-                                        //node.fixed = false;
-                                        visjsData.nodes.push(node);
-                                    }
-                                });
-                                visjsDataSource.edges.forEach(function (edge) {
-                                    if (!uniqueNodes[edge.id]) {
-                                        uniqueNodes[edge.id] = 1;
-                                        visjsData.edges.push(edge);
-                                    }
-                                });
+                        });
+                        visjsDataSource.edges.forEach(function (edge) {
+                            if (!uniqueNodes[edge.id]) {
+                                uniqueNodes[edge.id] = 1;
+                                visjsData.edges.push(edge);
                             }
-                            callbackEach();
                         });
                     }
+                    self.visjsData = null;
+                    callbackEach();
                 });
             },
             function (err) {
@@ -284,7 +274,7 @@ var KGquery_graph = (function () {
             },
         );
     };
-
+    */
     /**
      * Gets the implicit model data in Vis.js format.
      * @function
@@ -718,9 +708,16 @@ var KGquery_graph = (function () {
         var data_type = "KGmodelGraph";*/
         var fileName = KGquery.currentSource + "_KGmodelGraph.json";
         self.visjsData = null;
-        if (callback) {
-            options.callback = callback;
-        }
+
+        options.callback = function () {
+            self.visjsData = {};
+            self.visjsData.nodes = self.KGqueryGraph.data.nodes.get();
+            self.visjsData.edges = self.KGqueryGraph.data.edges.get();
+            if (callback) {
+                callback();
+            }
+        };
+
         self.KGqueryGraph.saveGraph(fileName, true, options);
 
         return;
@@ -984,7 +981,7 @@ var KGquery_graph = (function () {
             [
                 function (callbackSeries) {
                     KGquery_graph.message("generating tbox graph from abox graph");
-                    self.getImplicitModelVisjsData(KGquery.currentSource, function (err, result2) {
+                    self.getImplicitModelVisjsData(source, function (err, result2) {
                         if (err) {
                             return alert(err);
                         }
@@ -1141,8 +1138,8 @@ var KGquery_graph = (function () {
      *    - Node positions and scaling
      *    - Font settings
 
-     */
-    self.drawModel = function (displayGraphInList) {
+         */
+    self.drawModel = function (displayGraphInList, callback) {
         if (!self.visjsData) {
             return alert("no graph model");
         }
@@ -1225,6 +1222,9 @@ var KGquery_graph = (function () {
                 }
             });
             self.KGqueryGraph.data.nodes.update(nodes_fonts);
+            if (callback) {
+                callback();
+            }
         });
     };