+++ public/vocables/modules/tools/mappingModeler/dataSourcesManager.js
@@ -762,7 +762,7 @@ var DataSourceManager = (function () {
         var query =
             "PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>\n" +
             "PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>\n" +
-            "SELECT count(?sub) as ?triples ?table  " +
+            "SELECT (count(?sub) as ?triples) ?table  " +
             fromStr +
             " WHERE {\n" +
             "  ?sub ?pred ?obj . ?sub <http://souslesens.org/KGcreator#mappingFile> ?table\n" +