+++ public/vocables/modules/tools/mappingModeler/tripleFactory.js
@@ -337,7 +337,7 @@ var TripleFactory = (function () {
         var regex = /<([^>]*)> <*([^ >]*)>* <*([^>]*)>*/;
         data.sampleTriples.forEach(function (item, index) {
             var array = regex.exec(item);
-            //  tableData.push([array[1], array[2], array[3]]);
+            //tableData.push([array[1], array[2], array[3]]);
             tableData.push([escapeMarkup(array[1]), escapeMarkup(array[2]), escapeMarkup(array[3])]);
             //  tableData.push([escapeMarkup(item.s), escapeMarkup(item.p), escapeMarkup(item.o)]);
         });